Fix archive project through API
The `Projects::UpdateService` used by the API endpoint was duplicating
the checks done in `ArchiveContract` and `UnarchiveContract`. Now the
relevant contract is called and used instead. As the duplication has
been removed, the class `Projects::Archiver` is not needed anymore and
its methods have been inlined.

Also, as project is not active when being archived, all permission
checks would fail because no action is allowed on archived projects. So
the condition was relaxed a little to also allow permission check on
projects being archived.